New Downloader base class, which individual impl. need to extend
Downloader class takes care of the keyring, checking of the required crednetials are already configured (and will ask the user, if not), and HTTP requests.

Institution-specific implementations have to extend Downloader, and define the institution name, a unique code, the values required for logging in, and how to use those credentials to download files.